requirements.txt
python-dotenv
pytest
pytest-dotenv

jupyter
notion-client
openai

flask
flask_restplus
